fix: gif encode code check

+++ frameworks/kits/js/common/ndk/image_packer_native_impl.cpp
@@ -118,7 +118,7 @@ int32_t OH_ImagePackerNative::PackToDataMultiFrames(OHOS::Media::PackOption *opt
     if (ret != IMAGE_SUCCESS) {
         return ret;
     }
-    for (int i = 0; i < pixelmap.size(); i++) {
+    for (int i = 0; i < static_cast<int>(pixelmap.size()); i++) {
         ret = imagePacker->AddImage(*(pixelmap[i]->GetInnerPixelmap().get()));
     }
     if (ret != IMAGE_SUCCESS) {
@@ -222,7 +222,7 @@ int32_t OH_ImagePackerNative::PackToFileMultiFrames(OHOS::Media::PackOption *opt
     if (ret != IMAGE_SUCCESS) {
         return ret;
     }
-    for (int i = 0; i < pixelmap.size(); i++) {
+    for (int i = 0; i < static_cast<int>(pixelmap.size()); i++) {
         ret = imagePacker->AddImage(*(pixelmap[i]->GetInnerPixelmap().get()));
     }
     if (ret != IMAGE_SUCCESS) {
